progress on ticket 10
